pyproject.toml
[tool.poetry]
name = "board-game-finder"
version = "0.1.0"
description = "\"Our Project\""
authors = ["Erman Torun <etorun@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.7"
dbt-core = "^1.8.8"
dbt-bigquery = "^1.8.3"
apache-airflow = "^2.10.3"
apache-airflow-providers-postgres = "^5.14.0"
apache-airflow-providers-google = "^10.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
